Transcription verification (issue #14 )
- Changed option in "export as csv" to allow for blank spaces or empty
strings representing empty slots.
- Added option in "export as csv" for what do with the "ultrascrossed"
x-in-a-box symbol, which doesn't always show up properly in
spreadsheets. It can replaced with a symbol of the user's choice.
- Added "check transcription" button which verifies that the
transcription obeys whatever constraints are currently in place.
- Added "constraints" option to Settings menu. Users can select one of
two constraints at the moment (more can be added).
- Constraint 1 : no medial joints marked 'H'
- Constraint 2 : medial and distal joints must match in flexion
- Still to do: Add check for blank spaces and non-standard transcription
symbols (per issue #14)
- Still to do: Make sure constraint selection is persistent between
sessions. Currently, constraints are all reset to "off" when SLPA is
closed.